Removes filename argument from plot functions
Removes the `filename` argument from several plotting functions in the package.

The argument defaulted to NULL, resulting in in-memory processing, and was thus redundant as the user can directly save plots after they are generated.
This simplifies the function signatures and reduces redundancy.

Removes `stretch_raster_full` as this function is deprecated.